Attempted to prove "fold_right_max_returns_max".
I've taken this proof as far as I think I can without modifying the lemma
I've reduced it to something which is almost true: That the maximum of a list is in the list.
Unfortunately this is not quite true because we start at 0 rather than negative infinity...

It looks like this is where using the true identity of the max monoid with negative infinity
would help, because then the maximum of a list is always in the list.